Add Sanic Extension support
